chore(dotfiles): let Claude run `command rg`
Don't know why, but it tried to run `command rg` the other day instead
of `rg`. FWIW, I still see it prompting me for permission to run `rg`
sometimes. Like, if I say:

> Find files with "wincent" in the text. Prefer the fastest tool.

it will run it without prompting for confirmation.

But if I say:

> Use `rg` to find files with "wincent" in the text.

it sometimes prompts for confirmation. It's not exactly confidence
inspiring to see the arbitrariness with which these configs seem to
only apparently/intermittently work.