feat(nvim): add diffview.nvim plugin
Repo:

- https://github.com/sindrets/diffview.nvim

Adds minimal config to make it usable for basic use, like:

    :DiffviewOpen

to show current diff with file tree on left, and split diff (two panes)
on the right; or:

    :DiffviewFileHistory

to show, um, file history at bottom, and split diff (two panes) above.

Main things I've noticed that needed changing so far:

- Needed to suppress default mapping for `<tab>` because I use that for
  toggling folds. Using `gn` (ie. "[g]o [n]ext") instead.
- For symmetry, using `gp` (ie. "[g]o [p]revious") as inverse of `gn`.
- Suppressed icons warning.

Changed the `diff` entry in `'fillchars'` based on a suggestion from the
README; the box-drawing character tiles nicely and makes it look like
sections of the buffer are crossed out. I could also have used "╳" but
that's a bit too heavyweight, visually.

Also reworked highlight colors because it's basically impossible to see
what's going on otherwise. There is no word-diff in Neovim:

- https://github.com/neovim/neovim/issues/29549

But there is in Vim:

- https://github.com/vim/vim/pull/16881

Until that gets backported (which it might never be), I can either put
up with the very ugly highlighting I'm using now, or use a plugin like:

- https://github.com/rickhowe/diffchar.vim

I'll think about that. But in any case, in the meantime, starting with
the plug-in I'm adding in this commit. For more detailed info on usage,
see:

- https://github.com/sindrets/diffview.nvim/blob/main/USAGE.md

Contains useful examples of standard commands like:

    :DiffviewOpen origin/HEAD...HEAD

for showing the diff of current branch since the merge base.

Or, to review the same commit-by-commit:

    :DiffviewFileHistory --range=origin/HEAD...HEAD --right-only --no-merges

Or, to see the latest stash:

    :DiffviewFileHistory -g --range=stash

* aspects/nvim/files/.config/nvim/pack/bundle/opt/diffview.nvim 0000000...4516612 (312):
  > feat(actions): added `select_{prev|next}_commit` (#512)