feat(dotfiles): try out Claude hooks
- https://docs.anthropic.com/en/docs/claude-code/hooks

These taken from:

- https://github.com/schacon/cc-hooks-auto-commit

After reading:

- https://blog.gitbutler.com/automate-your-ai-workflows-with-claude-code-hooks/

First, I cloned the repo, and then I ran Claude inside it to test the
hooks. It raised errors, so I got it to "fix" them (it added a bunch of
`begin` blocks). I think all of this needs a pass before being declared
"final", but committing it anyway so I can kick the tires on it.