feat(nvim): set $RTP variable
Analogous to `$HELP` which I recently added. For example, I want to find
where `GetTypescriptIdent()` mentioned in the parent commit is defined,
but it's not mentioned in the help (can't find it with `:CommandTHelp`
or `:helpgrep GetTypescriptIdent`), and it's not in my installed plugins
either. In a case like this, can do `:Ack GetTypescriptIdent $RTP` and
it quickly leads me to:

/opt/homebrew/Cellar/neovim/0.11.4/share/nvim/runtime/indent/typescript.vim