feat(dotfiles): add Claude "Notification" hook
This will show a notification like:

   Claude
   Notification hook
   Claude needs your permission to use Bash

Which is useful if you have Command-Tabbed away and don't otherwise
notice that Claude is waiting for you.

According to the docs:

- https://docs.claude.com/en/docs/claude-code/hooks#notification

it should run in these two situations:

> 1. Claude needs your permission to use a tool. Example: “Claude
>    needs your permission to use Bash”
>
> 2. The prompt input has been idle for at least 60 seconds. “Claude
>    is waiting for your input”

The notification is in a sense redundant if Claude is in focus, but
given that I have my `terminal-notifier` notifications tuned to use
"banner" (ie. auto-hiding format), it's not too intrusive.

Speaking of `terminal-notifier`, I hadn't used it on this machine, so
had to grant permissions for it on first run. It didn't actually work
until I ran `killall NotificationCenter`.