refactor(nvim): store handles in .jsonc instead of in .json
I have enough handles in here now that I want to divide them into
sections and use comments as headers.

My first thought was to look up the help and see if
`vim.fn.json_decode()` could handle ".jsonc", at which point I
discovered there's a `vim.json.decode()` now. Lua is better than
Vimscript, so we switch.

Turns out that `readfile()` returns a list of lines, and
`vim.fn.json_decode()` handled that without transformation, but
`vim.json.decode()` wants a string.

This is fine though, because our goal was to support ".jsonc" after all,
and this gives us the perfect excuse to filter out comment lines before
joining back to a single string and passing it on.