chore: prepare to switch "vendor/git-cipher" from submodule to subtree
This commit removes the submodule; next commit will add the subtree.
Can't do both in one step because `git-subtree` insists on having a
clean worktree before starting.

In order to stay in a working state during this transition, I'm working
with a globally installed `git-cipher` (via `npm -g install git-cipher`)
and I've temporarily edited the path to `git-cipher` in `.git/config`
and in `.git/hooks/pre-commit`.