refactor: migrate away from `git subtree`
Well, that was short-lived. Discovered that running the updates in a
jj-colocated repo is a bit of a mess (have to stash, then run `git
subtree`, then run `jj git import` to integrate the changes).

Replacing it all with a dumb script that clones all repos under
`.cache/`, updates them, syncs over the results into place where they
can be committed, and emits a simple changelog (similar to what I had
with `git submodule summary`).