fix: turn regular file into a symlink
Not sure why this got committed as it was in the previous commit (it was
just a regular file with the word "encrypt" in it and no trailing
newline). Now, that _is_ how git models symlinks, but the mode was wrong
(ie. mode was 100644 instead of the desired 120000).

Whatever the cause (probably file it under "jj learning curve), both
`jj` and `git` agree that the file _is_ a symlink. Committing this
correctly now.