chore: commit Claude settings churn
This is a bit annoying, because every time I toggle thinking by hitting
the Tab key, it's going to flip that `alwaysThinkingEnabled` setting
(and note, in this case, it also reordered the sections within the
file). That feels like a fairly lightweight interaction to trigger
dotfile update. I'll separately look into how to keep this noise out of
the settings (and if that doesn't work, just get in the habit of leaving
"thinking" off and only ever turning it on temporarily).