feat(dotfiles): improve `tug` alias
As promised in the parent commit, let's explain what this all does now:

- If provided with more than one argument, print an error advising users
  to pass a single argument with "|", "&" or "~" operators.

- If provided with no arguments, behave as before (ie. find the closest
  pushable revision, and point the closest bookmark at it.

- Otherwise, the argument should specify the bookmark to move; point it
  at the closest pushable revision.

What does "closest pushable" mean?

- As before, we look at the `to` argument and all of its ancestors (ie.
  `::to`).

- We find the intersection with all mutable commits (ie. `& mutable()`);
  immutable commits are ones that we've already pushed to remotes, so
  this means unpushed commits.

- Additionally, we intersect with commits with non-empty descriptions
  (ie. `& ~description(exact:"")`); that is, we _exclude_ commits that
  have no descriptions.

- Likewise, we intersect with non-empty commits and merges.

- Again, we use `heads` to select the topmost commits (ie. without
  descendants) from this set.

Overall, this means that the "closest pushable" commit is the nearest
unpushed, non-empty commit, with a description.

The other piece here is that we allow an optional argument to `tug` in
case the use wants to override the default of `@`. In other words, with
no argument, `tug` will find the closest bookmark and move it to the
closest pushable commit, where "closest" is defined in relation to the
current commit. But if you pass an argument, it will move that bookmark
instead. Because you can pass an argument like `a | b`, you could
actually move more than one bookmark with this version of `tug`.

Finally, note that I've made a helper file that I can use from any
script-based alias moving forward. In this case, it is reading the color
values from `jj config` and using those to format its "Error:" and
"Hint:" messages.