chore: run `bin/update-dependencies` to sync with base16-nvim changes
Because my update process right now is:

1. Run `bin/update-dependencies` updates all dependencies. Commit all
   dependency updates.
2. Run `bin/update-themes` creates updated base16-nvim files (among
   other theme-related artifacts). Commit all theme-releated updates.
3. Commit the changes in `../base16-nvim` and push them.
4. Run `bin/update-dependencies` again (effectively updating the
   base16-nvim metadata only, because all the actual changes were
   included in the commit at step "2").

So this is a bit weird, but stems from a chicken-and-egg problem: you
can't update the base16-nvim files without updating the dependencies,
but when you update the dependencies and then create new base16-nvim
files, you wind up with a base16-nvim metadata update. Not sure what a
better approach would be, but for now, this is it.

================================================================================
CHANGELOG
================================================================================

aspects/nvim/files/.config/nvim/pack/bundle/opt/base16-nvim
  6e503dc..88a2f79

  88a2f79 (HEAD -> main, origin/main, origin/HEAD) chore: run theme update from dotfiles

================================================================================
1 dependency updated
================================================================================