refactor: switch from Yarn to npm
My use case is extremely simple: I have one runtime dependency and two
dev dependencies. Now that I am not vendoring Yarn any more, the benefit
of using it is marginal at best (maybe you could argue that `yarn
upgrade-interactive` justifies it), and not enough to offset the
headaches with corepack mentioned in the related issue. So, remove yarn
and just rely on npm instead. npm has come a long way since the dark old
days, so it's really not that bad, and by using it I avoid pulling yet
another dependency.

Closes: https://github.com/wincent/wincent/issues/163