perf(dotfiles): revert latency-minimization settings for kitty
When I accidentally write a bug that causes an infinite loop, or I
generate too much console output in Advent of Code, kitty can lock up in
a way that Ghostty does not. The necessary ingredient is flooding the
console with too much output. You can do this with something like:

    seq 1 1000000000000

This happens inside and outside of tmux. In both cases you see a beach
ball, and CTRL-C won't actually kill the process. Inside of tmux, you
can't switch panes or enter copy mode. Generally, the fix is to kill the
process that is producing the output.

Things that make it worse:

- Colorized output.
- Running in a vertical tmux split.

But, as I said, Ghostty remains responsive when this happens. I'm
reverting all three settings here, from the values set in 9d61b3fbafc1
("perf(dotfiles): apply latency-minimizing settings for kitty",
2024-12-27). Setting repaint_delay back to 10 (from 2) provides a
noticeable amount of mitigation, but it's setting sync_to_monitor back
to "yes" (from "no") that really fixes it, unambiguously.

I'd rather have a responsive-at-all terminal than one with a few ms less
response time in the happy path, so here goes the revert.

Change-Id: wvlmoyrsvrmnnyrntrosqovzwwuutmrs