refactor(dotfiles): template Docker config
Work wants to add some work-specic `"credHelpers"` config, but Docker
doesn't have any notion of "includes" that I could use to do this
separately on different machines.

So:

1. Ignore `~/.docker/config.json`
2. Instead, put base config in `~/.docker/config-base.json`
3. Put host-specific additions in `~/.docker/host/${HANDLE}.json`
4. Merge the host config (if any) into the base and template out the
   result.

This requires me to micromanage the .gitignore set-up to avoid
committing the generated config, and also to avoid committing the
plaintext version of the corp config, which is rather annoying and
fiddly.

Because I have to read in a file that might not be there, I had to do a
fairly ugly check-then-merge pattern, and also change the type of the
callback passed to the `variables()` function to allow it to be async
(ie. return a `Promise`).

This whole business is not very ergonomic, but I'm only doing it in one
place right now, so I don't want to prematurely abstract it. If I run
into a second place where this is necessary, I'll implement a more
streamlined pattern for it.

Change-Id: snxslkxlkrzvzqrtlptsswqyvqnlswmx