refactor: rename "violentmonkey" aspect to "userscripts"
Previously was called "tampermonkey". When I moved from Tampermonkey to
Violentmonkey (because I wanted to switch from a closed-source to an
open-source solution), I renamed it to "violentmonkey". But then Google
broke Violentmonkey (grrr), forcing me to switch back to Tampermonkey.

I don't want to keep renaming this thing. I was going to just call it
"browsermonkey" and be done with it, but given that I might one day
swtich to ScriptCat, monkey may be too specific. So, let's call it
"userscripts" instead.

Change-Id: rzqpunmtrmmqkvxyzoptpxwrvrtqzysw