feat(userscripts): add script to suppress PR previews in Slack
Based on a similar script that was shared at work, but instead of
just running every second, additionally runs eagerly on page load.

Appending "/s" to the URL stops Slack from showing a preview, but
doesn't break the URL (ie. GitHub dutifully serves up the PR page and,
critically, it doesn't redirect you back to the canonical URL without
the suffix).

I tried to find the origin of the "append /s" idea on the web, but my
search fu is failing me. For some other suffixes that I tried, I get
redirected straight back to the canonical URL. For some reason "/s"
sticks.

Change-Id: wrmrkqnkuozrukurvpsvwxoptwmsoqqn