fix: don't update dependency metadata when there are no actual changes
I noticed in f276a9b0673f5ddb9e3a713db7c7993e35976f23 that only one
dependency had actual changes, but we updated the metadata for all of
the dependencies that had had actual changes in the previous update
(f70a707b976d96c7e58ffe916f3dc43306be1cca).

To illustrate, look at what we do for yazi/vendor/plugins in the two
updates:

1.  Start first run.
2.  Update yazi/vendor/plugins from 19dc890 to 7ca66b4.
3.  Change `current` value for yazi/vendor/plugins in
    "dependencies.json" from 19dc890e33b8922eb1a3a165e685436ec4ac0a59
    to 7ca66b4103d76038a5d21c9b83766967f5028116.
4.  Start second run.
5.  Update nvim-lspconfig, but not yazi/vendor/plugins.
6.  Change `previous` value for yazi/vendor/plugins in
    "dependencies.json" from 19dc890e33b8922eb1a3a165e685436ec4ac0a59
    to 7ca66b4103d76038a5d21c9b83766967f5028116.

ie. in the first run we change `current` but not `previous`, and in the
second run we change `previous`.

The (untested) fix in this commit is to update both `current` and
`previous` only when they actual change. We'll see if this works in
future updates.

Change-Id: rqvnkpvoouquymtkknuwpwomuumkwzun