docs: add notes on terminal support
Found this reference:

- https://github.com/kovidgoyal/kitty/issues/8770#issuecomment-3026220182

indicating why kitty doesn't pass all of the tests in `bin/test-terminal`:

> kitty does not support invisible, overlined and blinking text. They
> are not considered worth the overhead.

But note that it may be a bit out of date because it was posted in
2025-07, and in my testing today, kitty _does_ support blinking text,
albeit only inside of tmux.

As for Sixel support:

>  I have no interest personally in implementing an inferior graphics
>  protocol in kitty. I know of no program that I would use that
>  supports sixels but not kitty's graphics protocol.

- https://github.com/kovidgoyal/kitty/issues/2511

Current status as of 2025-12-26:

    |                 | kitty  |        | Ghostty |         |
    |                 |   +    | kitty  |    +    | Ghostty |
    |                 |  tmux  |        |   tmux  |         |
    | --------------- | ------ | ------ | ------- | ------- |
    | Blinking text   |   ❌   |   ✅   |   ❌    |   ❌    |
    | Invisible text  |   ❌   |   ❌   |   ✅    |   ✅    |
    | Overline        |   ❌   |   ❌   |   ❌    |   ✅    |
    | Sixel graphics  |   ❌   |   ❌   |   ❌    |   ❌    |

None of these are really necessary IMO, and invisible text seems like an
antipattern, so probably no action recquired, although it is tempting to
see if I can at least get overline working in Ghostty + tmux, and
blinking working in kitty + tmux.

Change-Id: sstlsrptqrsosntmmnnmywtswzurlwur